requirements.txt
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
streamlit>=1.28.0
